requirements.txt
aniso8601==8.1.1
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.8
Flask==1.1.2
Flask-RESTful==0.3.8
google==3.0.0
google-api-core==1.25.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-core==1.5.0
google-cloud-translate==3.0.2
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
hatesonar==0.0.7
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
libcst==0.3.16
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
pandas==1.1.5
profanity==1.1
profanityfilter==2.0.6
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
scikit-learn==0.19.1
scipy==1.5.4
six==1.15.0
soupsieve==2.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
